+++ mod-data-export/src/test/resources/firebird/dataexport/features/export-authority-config-C432314.feature
@@ -1,4 +1,5 @@
 @parallel=false
+@C432314
 Feature: Verify configured limit of exported file size - Authorities (UUID)
 
   Background:
