+++ mod-data-export/src/test/resources/firebird/dataexport/features/export-for-holding-config-C432311.feature
@@ -1,3 +1,4 @@
+@C432311
 Feature: Verify configured limit of exported file size - Holdings (UUID)
 
   Background:
